requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
boto3==1.35.4
botocore==1.35.4
certifi==2024.7.4
charset-normalizer==3.3.2
distlib==0.3.8
Django==4.2.15
django-cors-headers==4.4.0
django-storages==1.14.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==23.0.0
idna==3.7
jmespath==1.0.1
packaging==24.1
pillow==10.4.0
platformdirs==4.2.2
psycopg2-binary==2.9.9
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-http-client==3.3.7
requests==2.32.3
s3transfer==0.10.2
sendgrid==6.11.0
six==1.16.0
sqlparse==0.5.1
starkbank-ecdsa==2.2.0
typing-extensions==4.12.2
urllib3==1.26.19
virtualenv==20.26.3
whitenoise==6.7.0
